+++ Changelog.md
@@ -3,7 +3,7 @@
 ## next
 date: 2013-12-22
 
-- simplify tests for command, add test for detecting full script path
+- improved tests
 
 ## 1.2.1
 date: 2013-12-22